# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/challenge/": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Creates challenge based in a given namespace",
                "parameters": [
                    {
                        "description": "Challenge",
                        "name": "namespace",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_endpoints.Challenge"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_endpoints.Challenge"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Deletes challenge in a namespace",
                "parameters": [
                    {
                        "description": "Challenge",
                        "name": "challenge",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_endpoints.DelChallenge"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_endpoints.DelRespChallenge"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api_endpoints.Challenge": {
            "type": "object",
            "properties": {
                "challengeName": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "ports": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "api_endpoints.DelChallenge": {
            "type": "object",
            "properties": {
                "challengeName": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
        "api_endpoints.DelRespChallenge": {
            "type": "object",
            "properties": {
                "challengeName": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
...
    }
}